+++ src/hooks/tradingView/index.ts
@@ -1,2 +1,4 @@
+export { useChartLines } from './useChartLines';
+export { useChartMarketAndResolution } from './useChartMarketAndResolution';
 export { useTradingView } from './useTradingView';
 export { useTradingViewTheme } from './useTradingViewTheme';
